# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TopMart ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicle-distribution/handoffs:
    get:
      operationId: listVehicleHandoffs
      tags: [vehicle-distribution]
      summary: List handoffs for the pilot vehicle
      description: >-
        Returns the prepared/lifecycle handoffs for the single active pilot
        vehicle (server-resolved). Authenticated via the dedicated
        vehicle-handoff auth wall (admin Bearer session OR the
        vehicle-distribution bot key). Fails closed (404) unless enabled, 503
        when enabled without schema approval.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleHandoffList"
        "401":
          description: Unauthorized
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "404":
          description: Feature disabled
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "503":
          description: Enabled but schema not approved
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    post:
      operationId: createVehicleHandoff
      tags: [vehicle-distribution]
      summary: Create a prepared handoff for the pilot vehicle
      description: >-
        Creates a prepared (not yet loaded) handoff. The actor, agent, vehicle
        and vehicle-warehouse target are all resolved server-side; only the
        source warehouse, line items, notes and a client operationKey are
        accepted. Idempotent on operationKey — replaying the same key returns
        the same handoff, while a payload mismatch on replay yields 409. No
        inventory is mutated.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleHandoffInput"
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/VehicleHandoffDetail" } } }
        "400":
          description: Invalid body
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401":
          description: Unauthorized
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "404":
          description: Feature disabled
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "409":
          description: Conflicting state or idempotency payload mismatch
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "503":
          description: Enabled but schema not approved
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  /vehicle-distribution/handoffs/{handoffId}:
    get:
      operationId: getVehicleHandoff
      tags: [vehicle-distribution]
      summary: Read a single handoff for the pilot vehicle
      parameters:
        - name: handoffId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/VehicleHandoffDetail" } } }
        "401":
          description: Unauthorized
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "404":
          description: Feature disabled or handoff not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "503":
          description: Enabled but schema not approved
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  /vehicle-distribution/handoffs/{handoffId}/labels/prepare:
    post:
      operationId: prepareVehicleHandoffLabels
      tags: [vehicle-distribution]
      summary: Prepare printable production-label passports for a prepared handoff
      description: >-
        Idempotently materialises one production_labels passport + one
        vehicle_label_claim + one label_prepared unit event per physical unit of
        a handoff in status 'prepared'. Requires PRODUCTION_LABELS_SCHEMA_APPROVED
        at request time (503 otherwise). Strict {operationKey} body. Barcodes are
        generated server-side (TS only). Replaying the same handoff + key +
        fingerprint returns the existing payload; a different key on the same
        handoff, the same key on another handoff, or a payload/fingerprint
        mismatch yields 409. No inventory is mutated.
      parameters:
        - name: handoffId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleHandoffOperationInput"
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/VehicleHandoffLabelsPayload" } } }
        "400":
          description: Invalid body
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401":
          description: Unauthorized
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "404":
          description: Feature disabled or handoff not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "409":
          description: Wrong state or idempotency conflict
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "503":
          description: Enabled but schema (vehicle or production labels) not approved
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  /vehicle-distribution/handoffs/{handoffId}/labels:
    get:
      operationId: getVehicleHandoffLabels
      tags: [vehicle-distribution]
      summary: Read the immutable printable label payload for a handoff
      description: >-
        Returns the deterministic, ordered print payload (persisted barcodes +
        all passport fields required by the PDF generator + print metadata) for a
        handoff whose labels have been prepared. Requires
        PRODUCTION_LABELS_SCHEMA_APPROVED at request time (503 otherwise).
      parameters:
        - name: handoffId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/VehicleHandoffLabelsPayload" } } }
        "401":
          description: Unauthorized
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "404":
          description: Feature disabled, handoff not found, or labels not prepared
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "503":
          description: Enabled but schema (vehicle or production labels) not approved
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  /vehicle-distribution/topmart-label-receipts:
    post:
      operationId: registerTopmartLabelReceipt
      tags: [vehicle-distribution]
      summary: Register immutable C-3 receipt provenance for sale labels
      description: >-
        Admin-only. Registers already-printed batch labels physically received
        into the C-3 destination immutably captured on one credited Top Mart
        sale (independent of later configuration). Requires exact authoritative SKU mapping and dona piece/weight
        coverage from that sale's immutable topmart-sale credit IN movements
        (not mutable catalog weight). The same sale and barcode set replays
        safely; a different set conflicts. Never mutates production_labels.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterTopmartLabelReceiptInput"
      responses:
        "200":
          description: Registered or idempotently replayed
          content: { application/json: { schema: { $ref: "#/components/schemas/TopmartLabelReceipt" } } }
        "400":
          description: Invalid body
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "403":
          description: Admin role required
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "404":
          description: Credited sale not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "409":
          description: Invalid labels, coverage, warehouse, or immutable-set conflict
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  /vehicle-distribution/handoffs/{handoffId}/labels/claim-existing:
    post:
      operationId: claimExistingVehicleHandoffLabels
      tags: [vehicle-distribution]
      summary: Claim already-printed production labels for a prepared C-3 handoff
      description: >-
        Claims existing, non-void printed batch labels without inserting or
        updating production_labels. The ordered barcode list must have exact SKU
        and piece coverage for every handoff item. Exact retries are idempotent;
        operation-key or payload conflicts return 409. On success claims and
        events are persisted and the handoff advances directly to labels_printed.
      parameters:
        - name: handoffId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimExistingVehicleHandoffLabelsInput"
      responses:
        "200":
          description: Existing labels claimed
          content: { application/json: { schema: { $ref: "#/components/schemas/VehicleHandoffLabelsPayload" } } }
        "400":
          description: Invalid body
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401":
          description: Unauthorized
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "404":
          description: Feature disabled, handoff or label not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "409":
          description: Wrong state, invalid label, coverage, or idempotency conflict
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "503":
          description: Enabled but schema not approved
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  /vehicle-distribution/handoffs/{handoffId}/confirm-labels-printed:
    post:
      operationId: confirmVehicleHandoffLabelsPrinted
      tags: [vehicle-distribution]
      summary: Confirm labels printed (first print or reprint) for a handoff
      description: >-
        Owns the full print-session contract. Requires
        PRODUCTION_LABELS_SCHEMA_APPROVED at request time (503 otherwise). Strict
        {operationKey} body. If a print session already exists for this
        operationKey on this handoff, returns the current payload with no
        increment. A first successful confirm on a 'prepared' handoff inserts a
        print session, marks each production_label printed (print_count +1,
        last_printed_at), advances claims prepared→printed, records idempotent
        label_printed events, and advances the handoff to labels_printed. A
        reprint confirm (allowed in labels_printed / handed_over /
        stock_transferred) inserts a reprint session and increments each passport
        once per NEW session, leaving handoff and loaded/sold claim states
        untouched. Reports isReprint and atLeastOnce (never claims exactly-once
        physical print). Cancelled/void/missing/mismatch yield 409.
      parameters:
        - name: handoffId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleHandoffOperationInput"
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/VehicleHandoffConfirmLabelsResult" } } }
        "400":
          description: Invalid body
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401":
          description: Unauthorized
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "404":
          description: Feature disabled or handoff not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "409":
          description: Wrong state or claim invariant not satisfied
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "503":
          description: Enabled but schema not approved
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
  /vehicle-distribution/handoffs/{handoffId}/handed-over:
    post:
      operationId: markVehicleHandoffHandedOver
      tags: [vehicle-distribution]
      summary: Transition labels_printed to handed_over
      description: >-
        Exact labels_printed → handed_over transition. Server actor/timestamp;
        no stock movement. Same-state retry is idempotent.
      parameters:
        - name: handoffId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleHandoffTransitionInput"
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/VehicleHandoffDetail" } } }
        "400":
          description: Invalid body
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401":
          description: Unauthorized
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "404":
          description: Feature disabled or handoff not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "409":
          description: Wrong state
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "503":
          description: Enabled but schema not approved
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
...
  /vehicle-distribution/handoffs/{handoffId}/cancel:
    post:
      operationId: cancelVehicleHandoff
      tags: [vehicle-distribution]
      summary: Cancel a pre-terminal handoff
      description: >-
        Cancels a handoff in any pre-terminal state (prepared, labels_printed,
        handed_over). Rejected once stock_transferred. No stock is mutated.
        Same cancelled-state retry is idempotent.
      parameters:
        - name: handoffId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleHandoffTransitionInput"
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/VehicleHandoffDetail" } } }
        "400":
          description: Invalid body
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "401":
          description: Unauthorized
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "404":
          description: Feature disabled or handoff not found
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "409":
          description: Already terminal (stock_transferred)
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
        "503":
          description: Enabled but schema not approved
          content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Vehicle Distribution (F2 pilot) ─────────────────────────────
...
    CreateVehicleHandoffInput:
      type: object
      description: >-
        Strict prepared-handoff creation body. Only these fields are accepted;
        agent, vehicle and vehicle-warehouse targets plus the actor are all
        resolved server-side. operationKey is a required client idempotency
        token.
      properties:
        sourceWarehouseId:
          type: integer
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/CreateVehicleHandoffItemInput"
        notes:
          type: ["string", "null"]
        operationKey:
          type: string
          minLength: 1
      required: [sourceWarehouseId, items, operationKey]
      additionalProperties: false
...
    VehicleHandoffList:
      type: object
      properties:
        vehicleId:
          type: integer
        handoffs:
          type: array
          items:
            $ref: "#/components/schemas/VehicleHandoffDetail"
      required: [vehicleId, handoffs]
    VehicleHandoffOperationInput:
      type: object
      description: >-
        Strict operation body for F4 label prepare/confirm. Carries only a
        required client idempotency operationKey; all other authority is
        resolved server-side.
      properties:
        operationKey:
          type: string
          minLength: 1
      required: [operationKey]
      additionalProperties: false
...
    RegisterTopmartLabelReceiptInput:
      type: object
      properties:
        saleId:
          type: integer
          minimum: 1
        barcodes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            pattern: "^TM[A-Z2-7]{16}$"
      required: [saleId, barcodes]
      additionalProperties: false
...